issue 408: reworked files and paths. Also removed a screen.debug() in a test that was being loud
